move APIPermissionError and PermissionOverwrite
