+++ app/models/channel.py
@@ -1,16 +1,11 @@
 from marshmallow import ValidationError
 from pymongo import ASCENDING
-from umongo import EmbeddedDocument, fields, validate
+from umongo import fields, validate
 
 from app.helpers.db_utils import instance
 from app.models.base import APIDocument
+from app.models.common import PermissionOverwrite
 from app.models.server import Server
-
-
-@instance.register
-class PermissionOverwrite(EmbeddedDocument):
-    role = fields.ReferenceField("Role")
-    permissions = fields.ListField(fields.StrField)
 
 
 @instance.register
