+++ app/models/common.py
@@ -1,0 +1,9 @@
+from umongo import EmbeddedDocument, fields
+
+from app.helpers.db_utils import instance
+
+
+@instance.register
+class PermissionOverwrite(EmbeddedDocument):
+    role = fields.ReferenceField("Role")
+    permissions = fields.ListField(fields.StrField)
