+++ app/models/section.py
@@ -2,6 +2,7 @@ from umongo import fields
 
 from app.helpers.db_utils import instance
 from app.models.base import APIDocument
+from app.models.common import PermissionOverwrite
 
 
 @instance.register
@@ -11,6 +12,8 @@ class Section(APIDocument):
     channels = fields.ListField(fields.ReferenceField("Channel"), default=[])
 
     position = fields.IntField()
+
+    permission_overwrites = fields.ListField(fields.EmbeddedField(PermissionOverwrite), default=[])
 
     class Meta:
         collection_name = "sections"
